Remove permission that breaks the tray icon
This permission allows us to communicate directly with the
org.kde.StatusNotifierWatcher interface, but this is actually a bad
idea.

Previously, when the app was about to set the tray icon, the
org.kde.StatusNotifierWatcher interface would receive the following
information:

$ dbus-monitor
...
   array [
      dict entry(
         string "Id"
         variant             string "cac90a20-a943-11ee-86a7-d9e05156d7eb"
      )
      dict entry(
         string "Category"
         variant             string "ApplicationStatus"
      )
      dict entry(
         string "Status"
         variant             string "Active"
      )
      dict entry(
         string "IconName"
         variant             string "/app/spotube/data/flutter_assets/assets/spotube-logo.png"
      )
...

The problem here is that the path for the icon file in the "IconName"
property does not exist outside of the Flatpak sandbox, and therefore
the icon file cannot be found or set.

By removing this permission, however, the current xdg-desktop-portal
implementation should take care of this problem for us automagically.

At least on KDE, instead of forwarding the literal path of the icon,
only the pixmap data of the icon is sent, through the "IconPixmap"
option:

$ dbus-monitor
...
   array [
      dict entry(
         string "Category"
         variant             string "ApplicationStatus"
      )
      dict entry(
         string "IconPixmap"
         variant             array [
               struct {
                  int32 16
                  int32 16
                  array of bytes [
                     ff 00 00 00 ff 00 00 00 ff 00 00 00 ff 00 00 00 ff 00 00
                     [ ... output shortened for readability ... ]
                  ]
               }
            ]
      )
...

Unfortunately, for reasons unbeknownst to me, simply removing the
org.kde.StatusNotifierWatcher permission is still not enough to make
the tray icon appear on Wayland sessions.

However, removing the Wayland permission to make the app run through
XWayland, did the trick.